Follow up rephrase from #512
Just rephrase the cast to float

+++ nipy/core/image/tests/test_image_spaces.py
@@ -23,7 +23,7 @@ from nose.tools import assert_true, assert_false, assert_equal, assert_raises
 
 def test_image_xyz_affine():
     # Test getting the image xyz affines
-    arr = np.arange(24).reshape((2,3,4))
+    arr = np.arange(24).reshape((2,3,4)).astype(float)
     aff = np.diag([2,3,4,1])
     img = Image(arr, vox2mni(aff))
     assert_true(is_xyz_affable(img))
@@ -34,7 +34,7 @@ def test_image_xyz_affine():
     img4_r = img4.reordered_axes([3,2,0,1])
     assert_false(is_xyz_affable(img4_r))
     assert_raises(AxesError, xyz_affine, img4_r)
-    nimg = nib.Nifti1Image(arr.astype('float'), aff)
+    nimg = nib.Nifti1Image(arr, aff)
     assert_true(is_xyz_affable(nimg))
     assert_array_equal(xyz_affine(nimg), aff)
     # Any dimensions not spatial, AxesError
@@ -77,7 +77,7 @@ def test_image_as_xyz_image():
     assert_array_equal(img.get_fdata(), img_t0_r.get_fdata())
     assert_equal(img.coordmap, img_t0_r.coordmap)
     # Test against nibabel image
-    nimg = nib.Nifti1Image(arr.astype('float'), np.diag([2,3,4,1]))
+    nimg = nib.Nifti1Image(arr, np.diag([2,3,4,1]))
     nimg_r = as_xyz_image(nimg)
     assert_true(nimg is nimg_r)
     # It's sometimes impossible to make an xyz affable image
